CAMEL-23530: Move DataWeave parser to camel-dataweave component for reuse
