+++ bom/camel-bom/pom.xml
@@ -709,6 +709,11 @@
       </dependency>
       <dependency>
         <groupId>org.apache.camel</groupId>
+        <artifactId>camel-dataweave</artifactId>
+        <version>4.22.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.camel</groupId>
         <artifactId>camel-debezium-common</artifactId>
         <version>4.22.0-SNAPSHOT</version>
       </dependency>
