+++ catalog/camel-allcomponents/pom.xml
@@ -599,6 +599,11 @@
         </dependency>
         <dependency>
             <groupId>org.apache.camel</groupId>
+            <artifactId>camel-dataweave</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.camel</groupId>
             <artifactId>camel-debezium-common</artifactId>
             <version>${project.version}</version>
         </dependency>
