+++ components/camel-dataweave/src/test/java/org/apache/camel/component/dataweave/DataWeaveLexerTest.java
@@ -14,23 +14,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.dsl.jbang.core.commands.transform;
+package org.apache.camel.component.dataweave;
 
 import java.util.List;
 
-import org.apache.camel.dsl.jbang.core.commands.transform.DataWeaveLexer.Token;
-import org.apache.camel.dsl.jbang.core.commands.transform.DataWeaveLexer.TokenType;
+import org.apache.camel.component.dataweave.DataWeaveLexer.Token;
+import org.apache.camel.component.dataweave.DataWeaveLexer.TokenType;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-/**
- * Unit tests for {@link DataWeaveLexer}.
- *
- * These assert the concrete token stream (types, values and positions) for representative inputs, focusing on the
- * tricky, ambiguous cases where a tokenizer typically goes wrong: multi-character operators, negative numbers vs the
- * minus operator, the {@code ---} header separator, and comment skipping.
- */
 class DataWeaveLexerTest {
 
     private static List<Token> tokenize(String input) {
@@ -50,7 +43,6 @@ class DataWeaveLexerTest {
 
     @Test
     void shouldClassifyKeywordsAndIdentifiers() {
-        // true/false -> BOOLEAN, null -> NULL_LIT, and/or/not -> logical ops, everything else -> IDENTIFIER
         assertEquals(
                 List.of(TokenType.BOOLEAN, TokenType.BOOLEAN, TokenType.NULL_LIT,
                         TokenType.AND, TokenType.OR, TokenType.NOT,
@@ -65,7 +57,6 @@ class DataWeaveLexerTest {
         assertEquals("hello", tokens.get(0).value());
         assertEquals(TokenType.STRING, tokens.get(1).type());
         assertEquals("world", tokens.get(1).value());
-        // the escaped quote is preserved verbatim inside the string value
         assertEquals("a\\\"b", tokens.get(2).value());
     }
 
@@ -78,7 +69,6 @@ class DataWeaveLexerTest {
 
     @Test
     void shouldTreatLeadingMinusAsNegativeNumberWhenNotAfterValue() {
-        // at the start of input the minus binds to the number
         List<Token> tokens = tokenize("-5");
         assertEquals(TokenType.NUMBER, tokens.get(0).type());
         assertEquals("-5", tokens.get(0).value());
@@ -86,7 +76,6 @@ class DataWeaveLexerTest {
 
     @Test
     void shouldTreatMinusBetweenValuesAsOperator() {
-        // "5 - 3": the minus follows a value, so it is the subtraction operator, not a sign
         assertEquals(
                 List.of(TokenType.NUMBER, TokenType.MINUS, TokenType.NUMBER, TokenType.EOF),
                 types("5 - 3"));
@@ -94,7 +83,6 @@ class DataWeaveLexerTest {
 
     @Test
     void shouldTreatMinusAfterOpeningParenAsNegativeNumber() {
-        // "(-3)": previous char is '(', not value-like, so the minus binds to the number
         assertEquals(
                 List.of(TokenType.LPAREN, TokenType.NUMBER, TokenType.RPAREN, TokenType.EOF),
                 types("(-3)"));
@@ -112,7 +100,6 @@ class DataWeaveLexerTest {
 
     @Test
     void shouldRecognizeHeaderSeparatorAtLineStart() {
-        // the --- on its own line is a header separator, while the negative number afterwards stays a number
         List<TokenType> ts = types("%dw 2.0\n---\n-5");
         assertEquals(
                 List.of(TokenType.PERCENT, TokenType.IDENTIFIER, TokenType.NUMBER,
@@ -133,7 +120,6 @@ class DataWeaveLexerTest {
 
     @Test
     void shouldTrackLineAndColumnPositions() {
-        // "ab\n  cd": first token at 1:1, second at line 2 after two spaces -> 2:3
         List<Token> tokens = tokenize("ab\n  cd");
         assertEquals(1, tokens.get(0).line());
         assertEquals(1, tokens.get(0).col());
@@ -145,7 +131,6 @@ class DataWeaveLexerTest {
 
     @Test
     void shouldSkipUnknownCharactersInsteadOfFailing() {
-        // '@' is not a recognized token; the lexer drops it and keeps going
         assertEquals(
                 List.of(TokenType.IDENTIFIER, TokenType.IDENTIFIER, TokenType.EOF),
                 types("a @ b"));
