+++ components/camel-dataweave/src/test/resources/dataweave/collection-map.dwl
