+++ components/camel-dataweave/src/test/resources/dataweave/event-message.dwl
