+++ components/camel-dataweave/src/test/resources/dataweave/null-handling.dwl
