+++ components/camel-dataweave/src/test/resources/dataweave/simple-rename.dwl
