+++ components/camel-dataweave/src/test/resources/dataweave/string-ops.dwl
