+++ components/camel-dataweave/src/test/resources/dataweave/type-coercion.dwl
