+++ components/pom.xml
@@ -120,6 +120,7 @@
         <module>camel-cyberark-vault</module>
         <module>camel-dapr</module>
         <module>camel-datasonnet</module>
+        <module>camel-dataweave</module>
         <module>camel-debug</module>
         <module>camel-diagram</module>
         <module>camel-debezium</module>
