+++ docs/components/modules/others/examples/json/dataweave.json
@@ -1,0 +1,1 @@
+../../../../../../components/camel-dataweave/src/generated/resources/dataweave.json
