+++ dsl/camel-jbang/camel-jbang-core/pom.xml
@@ -47,6 +47,10 @@
         <!-- camel -->
         <dependency>
             <groupId>org.apache.camel</groupId>
+            <artifactId>camel-dataweave</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.camel</groupId>
             <artifactId>camel-diagram</artifactId>
         </dependency>
         <dependency>
